requirements.txt
python-telegram-bot==20.7
youtube-transcript-api>=0.6.2
python-dotenv>=1.0.0
google-generativeai>=0.7.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
numpy>=1.24.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
pymongo>=4.6.0
python-multipart>=0.0.6